Add more tests for Asset Model index and store methods
